pyproject.toml
[tool.poetry]
name = "get-papers"
version = "0.1.0"
description = "Fetch PubMed papers with non-academic authors from pharma/biotech companies"
authors = ["Anjali Gupta anjali720gupta@gmail.com"]

[tool.poetry.dependencies]
python = ">=3.10"
biopython = "^1.83"
typer = "^0.12.3"
rich = "^13.7.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "cli.main:app"

